scripts/scraper/requirements.txt
# Web scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# Google Gemini AI
google-generativeai>=0.3.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0
